refactor: cleanup ragdoll playback
